% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create.baData.R
\name{create.baData}
\alias{create.baData}
\title{create \code{\link{ba}} object}
\usage{
create.baData(synbreedobj = NULL, pheno = NULL, geno = NULL, map = NULL,
  pedigree = NULL, family = NULL, covar = NULL, reorderMap = TRUE,
  map.unit = "cM", repeated = NULL, modCovar = NULL, A = NULL,
  Ainv = NULL, fixed = NULL, random = NULL, makeA = FALSE,
  makeAinv = FALSE)
}
\description{

}

