% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create.options.R
\name{create.options}
\alias{create.options}
\title{create \code{\link{op}} object}
\usage{
create.options(model = NULL, method = NULL, ante = NULL, poly = NULL,
  priors = NULL, init = NULL, D = "V", update_para = NULL,
  run_para = NULL, save.at = NULL, cv = NULL, print_mcmc = NULL,
  ncore = 1, seed = 1, convcrit = 1e-04)
}
\description{

}

