% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{createCV}
\alias{createCV}
\title{Quickly create cross-validation folds}
\usage{
createCV(formula, data, k = 5)
}
\arguments{
\item{formula}{formula of the trait name}

\item{data}{dataframe contain the phenotypes}

\item{k}{number of folds}
}
\value{
data.frame of the all traits including the cross-validation folds
}
\description{
Quickly create cross-validation folds
}
\examples{
\dontrun{
set.seed(1234)
PigPheno=createCV(data = PigPheno,k=5,"driploss")
}
}
