% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{man_plot_pvalue}
\alias{man_plot_pvalue}
\title{Create Manhattan plot for pvalue based models}
\usage{
man_plot_pvalue(ba, type = c("SNP", "Win"), col = c("black", "red"),
  ylim = c(0, 12), ...)
}
\arguments{
\item{ba}{an object of the class \code{ba} generated by baFit}

\item{type}{can be c("SNP","Win") for single SNP or window based Manhattan plot}

\item{col}{color for the Manhattan plot}

\item{ylim}{the range of yaxis}

\item{...}{other plot options}
}
\value{
Null. This function only prints to the screen.
}
\description{
Create Manhattan plot for pvalue based models
}
\examples{
\dontrun{
 rm(list=ls())
 library(BATools)
 data("Pig")
 #Standardize genotype matrix
 run_para=list(maxiter=100)
 update_para=list(df=FALSE,scale=TRUE,pi=FALSE)
 op<-create.options(model="GBLUP",method="REML",priors=NULL,init=init,
                  update_para=update_para,run_para=run_para,save.at="GBLUP",print_mcmc=NULL)
 gblup<-baFit(driploss~sex,data=PigPheno,geno=geno ,genoid = ~id,options = op,map=PigMap,GWA="SNP")
 par(mfrow=c(1,2))
 man_plot_pvalue(gblup)
 man_plot_pvalue(gblup,type="Win")
}
}
