% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PigPheno}
\alias{PigPheno}
\title{Pig Phenotypes}
\format{A data frame with 176 rows and 7 variables:
\describe{
  \item{test1}{test1}
  \item{test2}{test2}
  ...
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
PigPheno
}
\description{
A dataset containing the MSUPRP phenotypes
}
\keyword{datasets}
