% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baplot.R
\name{plot.ba}
\alias{plot.ba}
\title{Plot the results for the Bayesian models
 @title plot result for bafit}
\usage{
\method{plot}{ba}(BAout = NULL, type = c("pre", "trace"), op = NULL,
  iterStart = NULL, col = c("black", "red"))
}
\arguments{
\item{BAout}{\code{ba} object contains all the output of the Bayesian model}

\item{type}{\code{string} defines the type of the plot, which can be "pre","man","trace"
\tabular{ll}{
  \tab \code{"pre"}  plot the predicted value against the true value\cr
  \tab \code{"trace"}  Traceplot \cr
}}

\item{op}{\code{op} object created by create.options that is used for the analysis}

\item{iterStart}{the starting iteration for traceplot}
}
\description{
Plot the results for the Bayesian models
 @title plot result for bafit
}
\details{
You can either provide \code{baData} or y and Z for this function. If they're provided at the same time, \code{dataobj} will be used for the analysis
}
\examples{
baplot(dataobj=pig,BAout=ba,type="pre")
baplot(dataobj=pig,BAout=ba,type="trace",op=op)
}
