% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gwas.R
\name{postprob_plot}
\alias{postprob_plot}
\title{Manhattan plots of posterior probability}
\usage{
postprob_plot(postprob, map, col = c("black", "red"), chrom = NULL, ...)
}
\arguments{
\item{postprob}{A vector of posterior probability with SNP, marker or position names}

\item{map}{a genetic map: a data.frame with SNP or marker names in rows, chromosome name (chr column) and position (pos column)}

\item{col}{a vector of strings with color names for painting SNP according to their chromosome}

\item{...}{additional graphical parameters used by plot}

\item{threshold}{the threshold of significance}
}
\value{
NULL. A graphic is be generated
}
\description{
A function to create posterior probability plots from GWA and QTL mapping analyses.
}

