% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{print.ba}
\alias{print.ba}
\title{print \code{\link{ba}} object}
\usage{
\method{print}{ba}(ba)
}
\arguments{
\item{ba}{an object of the class gblup generated by \code{\link{ba}}}
}
\value{
Null. This function only prints to the screen:
}
\description{
A print method for ba objects
}
\seealso{
\code{\link{ba}}
}

