% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BATools.R
\name{bafit}
\alias{bafit}
\title{Fitting various Bayesian models}
\usage{
bafit(dataobj = NULL, op = NULL, y = NULL, Z = NULL, X = NULL,
  trait = NULL)
}
\arguments{
\item{dataobj}{A list of baData including phenotypes, genotypes and etc.}

\item{op}{A list of options to run Bayesian models}

\item{y}{A numeric vector of phenotypes}

\item{trait}{A string indicating the trait for analysis}

\item{A}{matrix of genotypes}
}
\value{
The result of the analysis as a ba object, a list of estimate of fixed and random effects as well as variance components.
}
\description{
bafit function can fit various Bayesian models including rrBLUP, BayesA/B/C, antedependence models and etc. Input data can be either a `baData` object or specify the fixed and random effects seperately.
}
\examples{
###########Loading and preparing data##########
library(BATools)
data("MSUPRP_sample")
summary(MSUPRP_sample)
pheno<-data.frame(MSUPRP_sample$pheno[,,])
geno<-MSUPRP_sample$geno[,1:500]
ped<-MSUPRP_sample$pedigree
map=MSUPRP_sample$map
sex<-ped$sex
sex<-as.factor(sex)
x<-model.matrix( ~ sex -1,contrasts.arg=list(sex=contrasts(sex, contrasts=F)))
colnames(x)<-c("female","male")
rownames(x)<-ped$ID
######################BayesA####################
##############Setting up options################
init=list(df=5,scale=0.01,pi=1)
run_para=list(niter=2000,burnIn=1000,skip=10)
print_mcmc=list(piter=200)
update_para=list(df=FALSE,scale=TRUE,pi=FALSE)
op<-create.options(model="BayesA",method="MCMC",ante=FALSE,priors=NULL,init=init,
                   update_para=update_para,run_para=run_para,save.at="BayesA",cv=NULL,print_mcmc=print_mcmc)
#################run BayesA MCMC################
ba<-bafit(dataobj=pig,op=op,trait="driploss")
ba
}

