% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BATools.R
\name{baFit}
\alias{baFit}
\title{Fitting various Bayesian models}
\usage{
baFit(formula, data, geno, genoid, randomFormula = NULL, map = NULL,
  PedAinv = NULL, options = NULL, train = NULL, GWA = c("No", "SNP",
  "Win"))
}
\arguments{
\item{formula}{A list of baData including phenotypes, genotypes and etc.}

\item{data}{A list of options to run Bayesian models}

\item{geno}{A numeric vector of phenotypes}

\item{genoid}{matrix of genotypes}

\item{randomFormula}{A string indicating the trait for analysis}
}
\value{
The result of the analysis as a `ba` object, a list of estimate of fixed and random effects as well as variance components.
}
\description{
baFit function can fit various Bayesian models including rrBLUP, BayesA/B/C, antedependence models and etc. Input data can be either a `baData` object or specify the fixed and random effects seperately.
}
\examples{
\dontrun{
#For GBLUP/rrBLUP:
demo(GBLUP)
demo(RR)

#For BayesA:
demo(BA)
demo(mapBA)

#For SSVS:
demo(SSVS)
demo(mapSSVS)

#For BayesB:
demo(BB)

}
}
