\name{BATools}
\alias{BATools}
\docType{package}
\title{
BATools
}
\description{
BATools is an R Package for Whole Genomic Analysis with Bayesian Models
}
\details{
\tabular{ll}{
Package: \tab BATools\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2015-09-22\cr
License: \tab GPL-3 \cr
}

}
\author{
Chunyu Chen, Wenzhao Yang, Heng Wang ,Robert J. Tempelman

Maintainer: Chunyu Chen <chench57@msu.edu>
}
\references{
W. Yang, R. J. Tempelman. 2012. A Bayesian Antedependence Model for Whole Genome Prediction. \emph{Genetics}, vol. 190 (4) pp. 1491-1501.
}




