\name{BATools}
\alias{BATools}
\docType{package}
\title{
BATools
}
\description{
BATools is an R Package for Whole Genomic Analysis with Bayesian Models
}
\details{
\tabular{ll}{
Package: \tab BATools\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2017-07-02\cr
License: \tab GPL-3 \cr
}

}
\author{
Chunyu Chen and Robert J. Tempelman

Maintainer: Chunyu Chen <chench57@msu.edu>
}
\references{
W. Yang, R. J. Tempelman. 2012. A Bayesian Antedependence Model for Whole Genome Prediction. \emph{Genetics}, vol. 190 (4) pp. 1491-1501.

C. Chen, R. J. Tempelman. 2015. An Integrated Approach to Empirical Bayesian Whole Genome Prediction Modeling. \emph{JABES}, Vol. 20 Issue 4, p491-511.

C. Chen, J. P. Steibel and R. J. Tempelman. 2017. Genome Wide Association Analyses Based on Broadly Different Specifications for Prior Distributions, Genomic Windows, and Estimation Methods.\emph{Genetics}, https://doi.org/10.1534/genetics.117.202259.


}




