% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BATools.R
\docType{package}
\name{BATools}
\alias{BATools}
\alias{BATools-package}
\title{BATools}
\description{
BATools: An R package that extends 'Bayesian Alphabet' for Whole Genome Prediction and Genome-wide association study
}
\details{
\describe{
 Package: BATools
 
 Version: 1.1.3
 
 Date: 2017-07-04
 
 License: GPL-3
}
}
\references{
W. Yang, R. J. Tempelman. 2012. A Bayesian Antedependence Model for Whole Genome Prediction. \emph{Genetics}, vol. 190 (4) pp. 1491-1501.

C. Chen, R. J. Tempelman. 2015. An Integrated Approach to Empirical Bayesian Whole Genome Prediction Modeling. \emph{JABES}, Vol. 20 Issue 4, p491-511.

W. Yang, C. Chen, R. J. Tempelman. 2015. Improving the computational efficiency of fully Bayes inference and assessing the effect of misspecification of hyperparameters in whole-genome prediction models. \emph{Genetics Selection Evolution}, Vol. 47	Issue 1.

C. Chen, J. P. Steibel and R. J. Tempelman. 2017. Genome Wide Association Analyses Based on Broadly Different Specifications for Prior Distributions, Genomic Windows, and Estimation Methods.\emph{Genetics}, https://doi.org/10.1534/genetics.117.202259.
}
\seealso{
\code{\link{baFit}} for fitting the model
}
\author{
Chunyu Chen and Robert J. Tempelman
 
 Maintainer: Chunyu Chen <chench57@msu.edu>
}
