% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create.baData.R
\name{create.baData}
\alias{create.baData}
\title{create \code{\link{ba}} object}
\usage{
create.baData(synbreedobj = NULL, pheno = NULL, geno = NULL, map = NULL,
  pedigree = NULL, family = NULL, covar = NULL, reorderMap = TRUE,
  map.unit = "cM", repeated = NULL, modCovar = NULL, A = NULL,
  Ainv = NULL, fixed = NULL, random = NULL, makeA = FALSE,
  makeAinv = FALSE)
}
\arguments{
\item{synbreedobj}{A pre-created synbreed data object}

\item{pheno}{\code{data.frame} with individuals organized in rows and traits organized in columns. For unrepeated measures unique rownames should identify individuals. For repeated measures, the first column identifies individuals and a second column indicates repetitions (see also argument \code{repeated}).}

\item{geno}{\code{matrix} with individuals organized in rows and markers organized in columns. Genotypes could be coded arbitrarily. Missing values should be coded as \code{NA}. Colums or rows with only missing values not allowed. Unique \code{rownames} identify individuals and unique \code{colnames} markers. If no \code{rownames} are available, they are taken from element \code{pheno} (if available and if dimension matches).  If no \code{colnames} are used, the \code{rownames} of \code{map} are used if dimension matches.}

\item{map}{\code{data.frame} with one row for each marker and two columns (named \code{chr} and \code{pos}). First columns gives the chromosome (\code{numeric} or \code{character} but not \code{factor}) and second column the position  on the chromosome in centimorgan or the physical distance relative to the reference sequence in basepairs. Unique \code{rownames} indicate the marker names which should match with marker names in \code{geno}. Note that order and number of markers must not be identical with the order in \code{geno}. If this is the case, gaps in the map are filled with \code{NA} to ensure the same number and order as in element \code{geno} of the resulting \code{gpData} object.}

\item{pedigree}{Object of class \code{pedigree}.}

\item{family}{\code{data.frame} assigning individuals to families with names of individuals in \code{rownames} This information could be used for replacing of missing values with function \code{codeGeno} in synbreed.}

\item{covar}{\code{data.frame} with further covariates for all individuals that either appear in \code{pheno}, \code{geno} or \code{pedigree$ID},  e.g. sex or age. \code{rownames} must be specified to identify individuals. Typically this element is not specified by the user.}

\item{reorderMap}{\code{logical}. Should markers in \code{geno} and \code{map} be reordered by chromosome number and position within chromosome according to \code{map} (default = \code{TRUE})?}

\item{map.unit}{Character. Unit of position in \code{map}, i.e. 'cM' for genetic distance or 'bp' for physical distance (default = 'cM').}

\item{repeated}{This column is used to identify the replications of the phenotypic values. The unique values become the names of the third dimension of the pheno object in the \code{gpData}. This argument is only required for repeated measurements.}

\item{modCovar}{\code{vector} with \code{colnames} which identify columns with covariables in \code{pheno}. This argument is only required for repeated measurements.}

\item{A}{\code{matrix} of additive genetic relationship matrix}

\item{Ainv}{\code{matrix} of inverse of A}

\item{fixed}{\code{matrix} or \code{factor} of fixed effects}

\item{random}{\code{matrix} or \code{factor} of random effects}

\item{makeA}{\code{logical} indicating if create A matrix based on the pedigree information}

\item{makeAinv}{\code{logical} indicating if create A inverse matrix based on the pedigree information, A will be created before creating Ainv is created}
}
\description{
create \code{\link{ba}} object
}
\details{
The object \code{baData} is designed to provide a unified data object for whole genome prediction. It inherits the \code{gpData} object from synbreed package and is extended with more optional data item for whole genome prediction. So \code{baData} can be used in any functions in synbreed package including data imputation and etc.
}
\examples{
#####create baData from gpData######
data("MSUPRP_sample") #load Michigan State University pig data
pheno<-data.frame(MSUPRP_sample$pheno[,,])
geno<-MSUPRP_sample$geno[,1:500]
ped<-MSUPRP_sample$pedigree
map=MSUPRP_sample$map
sex<-ped$sex
sex<-as.factor(sex)
x<-model.matrix( ~ sex -1,contrasts.arg=list(sex=contrasts(sex, contrasts=F)))
colnames(x)<-c("female","male")
rownames(x)<-ped$ID
#create baData object using synbreed object
pig.syn<-create.baData(synbreedobj=MSUPRP_sample,fixed=x,makeAinv=T)
names(pig.syn)
#create baData object using raw data
pig=create.baData(pheno=pheno,geno=geno,map=map,pedigree=ped,fixed=x,makeAinv=F)
}

