% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{createCV}
\alias{createCV}
\title{Quickly create cross-validation folds}
\usage{
createCV(data, k = 5, y)
}
\arguments{
\item{data}{dataframe contain the phenotypes}

\item{k}{number of folds}

\item{y}{charactor of the trait name}
}
\value{
data.frame of the all traits including the cross-validation folds
}
\description{
Quickly create cross-validation folds
}
