% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{print.ba}
\alias{print.ba}
\title{A print method for ba objects}
\usage{
\method{print}{ba}(ba)
}
\arguments{
\item{ba}{an object of the class ba generated by bafit}
}
\value{
Null. This function only prints to the screen.
}
\description{
A print method for ba objects
}
