% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{print.ba}
\alias{print.ba}
\title{A print method for ba objects}
\usage{
print.ba(x, ...)
}
\arguments{
\item{x}{an object of the class \code{ba} generated by bafit}
}
\value{
Null. This function only prints to the screen.
}
\description{
A print method for ba objects
}
\examples{
\dontrun{
 rm(list=ls())
 library(BATools)
 data("Pig")
 #Standardize genotype matrix
 #or set your own starting values using 
 #init=list(df=5,scale=0.01,pi=1) 
 run_para=list(niter=2000,burnIn=1000,skip=10)
 print_mcmc=list(piter=500)
 update_para=list(df=FALSE,scale=TRUE,pi=F)
 op<-create.options(model="SSVS",method="MCMC",seed=1,priors=NULL,init=init,
                  update_para=update_para,run_para=run_para,save.at="SSVS",print_mcmc=print_mcmc)
 SSVS<-baFit(driploss~sex,data=PigPheno,geno=geno ,genoid = ~id,options = op,map=PigMap,GWA="Win")
 SSVS
}
}
