% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{set_init}
\alias{set_init}
\alias{set_init.baData}
\alias{set_init.default}
\title{Set initial values for hyperparameters}
\usage{
set_init(y, ...)

\method{set_init}{default}(y, Z, df = 5, scale = NULL, vare = NULL,
  pi_snp = 0.05, h2 = 0.5, c = 1000, model = "rrBLUP", centered = T)

\method{set_init}{baData}(y, df = 5, scale = NULL, vare = NULL,
  pi_snp = 0.05, h2 = 0.5, c = 1000, model = "rrBLUP", centered = T,
  trait = 1)
}
\arguments{
\item{y}{`numeric` of phenotypes or `baData` object}

\item{Z}{genotype matrix}

\item{df}{degrees of freedom parameter}

\item{scale}{scale parameter, if is `NULL` set automatically}

\item{vare}{residual variance, if is `NULL` set automatically}

\item{pi_snp}{proportion of SNP that have non-zero effect for BayesB and SSVS}

\item{h2}{hertiability}

\item{c}{ratio for the smaller variance}

\item{model}{the model used for analysis, can be 'rrBLUP', 'BayesA','BayesB', or "SSVS"}

\item{centered}{logical indicating if Z is centered}

\item{baobj}{`baData` object}
}
\value{
a list of initial values
}
\description{
Set initial values for hyperparameters
}

