% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{set_init}
\alias{set_init}
\title{Set initial values for hyperparameters}
\usage{
set_init(y, data, geno, genoid, df = 5, scale = NULL, vare = NULL,
  g = NULL, beta = NULL, pi_snp = 0.05, h2 = 0.5, c = 1000,
  model = c("rrBLUP", "GBLUP", "BayesA", "BayesB", "SSVS"), centered = T,
  from = NULL)
}
\arguments{
\item{y}{character phenotypes name}

\item{data}{data.frame containing phenotypes}

\item{geno}{genotype matrix}

\item{genoid}{character of the column name of `data` representing the id related the rowname of `geno`}

\item{df}{degrees of freedom parameter}

\item{scale}{scale parameter, if is `NULL` set automatically}

\item{vare}{residual variance, if is `NULL` set automatically}

\item{g}{starting values for SNP effects, if is `NULL` will start from}

\item{beta}{starting values for fixed effects,}

\item{pi_snp}{proportion of SNP that have non-zero effect for BayesB and SSVS, default is 0.05}

\item{h2}{hertiability}

\item{c}{ratio for the smaller variance for SSVS}

\item{model}{the model used for analysis, can be 'rrBLUP',"GBLUP", 'BayesA','BayesB', or "SSVS", for single step or antedependence model use the base model}

\item{centered}{logical indicating if Z is centered}

\item{from}{the source of scale, if it's 'rrBLUP', the scale will be adjusted for BayesA and SSVS}
}
\value{
a list of initial values
}
\description{
Set initial values for hyperparameters
}
