% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{std_geno}
\alias{std_geno}
\title{Re-center genotype matrix}
\usage{
std_geno(x, method = c("s", "c"), freq = NULL)
}
\arguments{
\item{x}{genotype matrix or `baData` object}

\item{method}{the method used for centering, `c` is centering and `s` is standardizing}
}
\value{
centered genotype matrix  or `baData` object with centered genotype matrix
}
\description{
Re-center genotype matrix
}
