% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{std_geno}
\alias{std_geno}
\alias{std_geno.baData}
\alias{std_geno.default}
\title{Re-center genotype matrix}
\usage{
std_geno(x, ...)

\method{std_geno}{default}(x, method = "s")

\method{std_geno}{baData}(x, method = "s")
}
\arguments{
\item{x}{genotype matrix}

\item{method}{the method used for centering, `c` is centering and `s` is standardizing}

\item{baobj}{`baData` object}
}
\value{
centered genotype matrix  or `baData` object with centered genotype matrix
}
\description{
Re-center genotype matrix
}

